refactor: 빌더 패턴, 예외처리 통일
