+++ src/main/java/xyz/tomorrowlearncamp/newsfeed/domain/commentlike/controller/CommentLikeController.java
@@ -1,12 +1,11 @@
-package xyz.tomorrowlearncamp.newsfeed.domain.comment_like.controller;
+package xyz.tomorrowlearncamp.newsfeed.domain.commentlike.controller;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import xyz.tomorrowlearncamp.newsfeed.auth.dto.LoginUserResponseDto;
-import xyz.tomorrowlearncamp.newsfeed.domain.comment_like.dto.CommentLikeRequestDto;
-import xyz.tomorrowlearncamp.newsfeed.domain.comment_like.service.CommentLikeService;
+import xyz.tomorrowlearncamp.newsfeed.domain.commentlike.service.CommentLikeService;
 import xyz.tomorrowlearncamp.newsfeed.global.etc.Const;
 
 @RestController
@@ -17,10 +16,10 @@ public class CommentLikeController {
     private final CommentLikeService commentLikeService;
 
     @PostMapping
-    public ResponseEntity<String> toggleLike(@PathVariable Long commentsId,
+    public ResponseEntity<Void> toggleLike(@PathVariable Long commentsId,
                                              @SessionAttribute(name = Const.LOGIN_USER) LoginUserResponseDto responseDto) {
-        boolean isLiked = commentLikeService.toggleLike(responseDto.getId(), commentsId);
-        return new ResponseEntity<>(isLiked ? "좋아요" : "좋아요 취소", HttpStatus.OK);
+        commentLikeService.toggleLike(responseDto.getId(), commentsId);
+        return new ResponseEntity<>(HttpStatus.OK);
     }
 
 }
