+++ src/main/java/xyz/tomorrowlearncamp/newsfeed/domain/commentlike/repository/CommentLikeRepository.java
@@ -1,8 +1,8 @@
-package xyz.tomorrowlearncamp.newsfeed.domain.comment_like.repository;
+package xyz.tomorrowlearncamp.newsfeed.domain.commentlike.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import xyz.tomorrowlearncamp.newsfeed.domain.comment.entity.Comment;
-import xyz.tomorrowlearncamp.newsfeed.domain.comment_like.Entity.CommentLike;
+import xyz.tomorrowlearncamp.newsfeed.domain.commentlike.Entity.CommentLike;
 import xyz.tomorrowlearncamp.newsfeed.domain.user.entity.Users;
